requirements.txt
asgiref==3.5.2
dj-database-url==2.1.0
Django==4.1.11
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
inflection==0.5.1
python-decouple==3.8
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.7.1
tzdata==2022.6
